refactor(src):Lotto creation methods inside LottoBuilder class go to LottoList class
LottoList 클래스 생성하여 LottoBuilder안에 있는 로또 생성 메서드들 LottoList 클래스로 이동
